@@ -26,158 +24,122 @@
     "build:appimage-arm64": "node scripts/build.mjs appimage-arm64",
     "build:windows": "node scripts/build.mjs windows",
     "build:dir": "pnpm clean:build && pnpm bundle && electron-builder --dir",
-    "bundle": "concurrently 'pnpm bundle:main' 'pnpm bundle:preload' 'pnpm bundle:renderer'",
-    "bundle:main": "cross-env NODE_ENV=production webpack --config ./webpack/main.config.js --color",
-    "bundle:preload": "cross-env NODE_ENV=production webpack --config ./webpack/preload.config.js --color",
-    "bundle:renderer": "cross-env NODE_ENV=production webpack --config ./webpack/renderer.config.js --color",
-    "start": "cross-env NODE_ENV=production electron ./bundle/main.prod.js",
-    "dev:bundle:main": "cross-env NODE_ENV=development webpack --config ./webpack/main.config.js --color",
-    "dev:bundle:preload": "cross-env NODE_ENV=development webpack --config ./webpack/preload.config.js --color",
-    "dev:start": "cross-env NODE_ENV=development electron -r @babel/register ./bundle/main.dev.js",
-    "dev:start-server": "cross-env NODE_ENV=development webpack serve --hot --config ./webpack/renderer.config.js",
+    "start": "pnpm dev:html && electron-vite preview",
+    "dev": "pnpm dev:html && electron-vite dev --watch",
+    "dev:html": "cross-env NODE_ENV=development ejs ./src/index.ejs -o ./src/renderer/index.html",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "lint": "concurrently 'pnpm lint:code' 'pnpm lint:styles'",
-    "lint:code": "eslint '**/*.[j|t]s{,x}'",
+    "lint:code": "ESLINT_USE_FLAT_CONFIG=true eslint \"**/*.{j,mj,cj,t}s\"",
     "lint:code:fix": "pnpm lint:code --fix",
     "lint:styles": "stylelint --ignore-path .eslintignore '**/*.pcss'",
     "lint:styles:fix": "stylelint --ignore-path .eslintignore '**/*.pcss' --fix",
-    "test": "pnpm test:coverage && pnpm format && pnpm lint",
+    "test": "pnpm test:unit && pnpm format && pnpm lint",
     "test:e2e": "pnpm build:dir && pnpx wdio ./wdio.conf.js",
-    "test:coverage": "jest --coverage",
-    "test:watch": "jest --watch",
-    "test:watch-coverage": "jest --watch --coverage",
+    "test:unit": "vitest --coverage --watch=false",
+    "test:dev": "vitest --coverage",
     "release": "node scripts/create-draft-release.mjs",
-    "prepare": "husky install"
+    "prepare": "husky"
   },
   "dependencies": {
-    "@goosewobbler/electron-redux": "^1.0.4"
+    "bufferutil": "^4.0.8",
+    "utf-8-validate": "^6.0.3",
+    "zustand": "^4.5.2",
+    "zutron": "^0.0.3"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.0",
-    "@babel/core": "^7.17.2",
-    "@babel/plugin-proposal-optional-chaining": "^7.16.7",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@babel/register": "^7.17.0",
-    "@headlessui/react": "^1.4.3",
-    "@heroicons/react": "^1.0.5",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
-    "@redux-offline/redux-offline": "^2.6.0",
-    "@reduxjs/toolkit": "^1.7.2",
-    "@testing-library/dom": "^8.11.3",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@testing-library/webdriverio": "^3.0.5",
-    "@types/jest": "^27.4.0",
-    "@types/mocha": "^9.1.0",
-    "@types/node": "^17.0.17",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.22",
-    "@types/react-tabs": "^2.3.4",
-    "@types/request": "^2.48.8",
-    "@types/source-map-support": "^0.5.2",
-    "@types/tailwindcss": "^3.0.7",
-    "@types/webpack": "^5.28.0",
-    "@types/webpack-env": "^1.16.3",
-    "@typescript-eslint/eslint-plugin": "^5.11.0",
-    "@typescript-eslint/parser": "^5.11.0",
-    "@wdio/cli": "^7.16.15",
-    "@wdio/local-runner": "^7.16.15",
-    "@wdio/mocha-framework": "^7.16.15",
-    "autoprefixer": "^10.4.2",
-    "babel-loader": "^8.2.3",
-    "ci-info": "^3.3.0",
-    "classnames": "^2.2.6",
-    "concurrently": "^7.0.0",
-    "copy-webpack-plugin": "^10.2.4",
+    "@electron/notarize": "^2.3.0",
+    "@headlessui/react": "^1.7.18",
+    "@heroicons/react": "^2.1.1",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@testing-library/dom": "^10.1.0",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^15.0.6",
+    "@testing-library/user-event": "^14.5.2",
+    "@testing-library/webdriverio": "^3.2.1",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.12.5",
+    "@types/react": "^18.2.57",
+    "@types/react-dom": "^18.2.19",
+    "@types/request": "^2.48.12",
+    "@types/source-map-support": "^0.5.10",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
+    "@vitejs/plugin-react-swc": "^3.6.0",
+    "@vitest/coverage-v8": "^1.3.1",
+    "@wdio/cli": "^8.32.2",
+    "@wdio/local-runner": "^8.32.2",
+    "@wdio/mocha-framework": "^8.32.2",
+    "@wdio/types": "^8.32.2",
+    "autoprefixer": "^10.4.19",
+    "ci-info": "^4.0.0",
+    "classnames": "^2.5.1",
+    "concurrently": "^8.2.2",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.6.0",
-    "cssnano": "^5.0.17",
-    "dnd-core": "^15.1.1",
-    "electron": "^17.0.0",
-    "electron-builder": "^22.14.13",
-    "electron-debug": "^3.2.0",
-    "electron-devtools-installer": "^3.2.0",
-    "electron-notarize": "^1.1.1",
-    "electron-store": "^8.0.1",
-    "eslint": "^8.9.0",
-    "eslint-config-airbnb": "^19.0.3",
-    "eslint-config-airbnb-typescript": "^16.1.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-typescript": "^3.0.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-import-resolver-webpack": "^0.13.2",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest": "^26.1.0",
-    "eslint-plugin-jest-dom": "^4.0.1",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "eslint-plugin-testing-library": "^5.0.5",
-    "eslint-plugin-webdriverio": "^1.0.1",
-    "expect-webdriverio": "^3.1.4",
+    "css-loader": "^7.0.0",
+    "cssnano": "^7.0.1",
+    "dnd-core": "^16.0.1",
+    "ejs": "^3.1.9",
+    "electron": "^30.0.2",
+    "electron-builder": "^24.12.0",
+    "electron-debug": "^4.0.0",
+    "electron-devtools-assembler": "^1.2.0",
+    "electron-store": "^9.0.0",
+    "electron-vite": "^2.0.0",
+    "esbuild-style-plugin": "^1.6.3",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-vitest": "^0.4.1",
+    "eslint-plugin-wdio": "^8.24.12",
+    "expect-webdriverio": "^4.11.9",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^7.2.1",
-    "html-webpack-harddisk-plugin": "^2.0.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^7.0.0",
-    "jest": "^27.5.1",
-    "lint-staged": "^12.3.3",
-    "mini-css-extract-plugin": "^2.5.3",
+    "globals": "^15.0.0",
+    "husky": "^9.0.11",
+    "jsdom": "^24.0.0",
+    "lint-staged": "^15.2.2",
     "node-loader": "^2.0.0",
-    "postcss": "^8.4.6",
-    "postcss-cli": "^9.1.0",
-    "postcss-loader": "^6.2.1",
-    "prettier": "^2.5.1",
-    "prettier-plugin-tailwindcss": "^0.1.7",
-    "react": "^17.0.2",
-    "react-dnd": "^15.1.1",
-    "react-dnd-html5-backend": "^15.1.2",
-    "react-dom": "^17.0.2",
+    "postcss": "^8.4.38",
+    "postcss-cli": "^11.0.0",
+    "postcss-loader": "^8.1.0",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "react": "^18.2.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
+    "react-dom": "^18.2.0",
     "react-multi-split-pane": "^0.3.3",
-    "react-redux": "^7.2.6",
-    "react-refresh": "^0.11.0",
-    "react-tabs": "^3.2.3",
-    "redux": "^4.1.2",
-    "source-map-loader": "^3.0.1",
+    "react-refresh": "^0.14.0",
+    "react-tabs": "^6.0.2",
+    "read-package-up": "^11.0.0",
+    "source-map-loader": "^5.0.0",
     "source-map-support": "^0.5.21",
-    "stylelint": "^14.5.0",
-    "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-standard": "^25.0.0",
-    "tailwindcss": "^3.0.22",
+    "stylelint": "^16.2.1",
+    "stylelint-config-standard": "^36.0.0",
+    "tailwindcss": "^3.4.3",
     "tailwindcss-scoped-groups": "^2.0.0",
-    "terser-webpack-plugin": "^5.3.1",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^10.5.0",
-    "typescript": "^4.5.5",
-    "wdio-electron-service": "^2.0.5",
-    "webdriverio": "^7.16.15",
-    "webpack": "^5.68.0",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
-    "webpack-electron-reload": "^1.0.1"
+    "ts-node": "^10.9.2",
+    "tsup": "^8.0.2",
+    "typescript": "^5.4.4",
+    "vite": "^5.1.4",
+    "vitest": "^1.3.1",
+    "wdio-electron-service": "^6.3.1",
+    "webdriverio": "^8.32.2"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=20.9.0"
   },
   "lint-staged": {
-    "**/*.{js,ts}": [
-      "eslint --fix"
+    "**/*.{mjs,js,ts,tsx}": [
+      "pnpm lint:code"
     ],
-    "**/*.{yml,json,md,ts,js}": [
+    "**/*.{yml,json,md,ts,tsx,js,mjs}": [
       "prettier --write"
     ]
   },
   "browserslist": [
-    "electron >= 17",
-    "node >= 18"
+    "electron >= 29"
   ]
 }
