@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const { execSync } = require('child_process');
-const { notarize } = require('electron-notarize');
+const { notarize } = require('@electron/notarize');
 
 module.exports = async (context) => {
   const {
