@@ -1,7 +1,8 @@
 import React, { ReactElement } from 'react';
-import { Browser, LoadContextType } from '../../common/types';
-import { log } from '../../main/helpers/console';
-import { BrowserTrack } from '../tracks/BrowserTrack';
+
+import { log } from '../../main/helpers/console.js';
+import { BrowserTrack } from '../tracks/BrowserTrack.js';
+import { Browser, LoadContextType } from '../../common/types.js';
 
 export function MetaPanel({ browser: { id, tracks } }: { browser: Browser }): ReactElement {
   log('tracks', tracks);
