@@ -1,17 +1,13 @@
-import { BrowserView, BrowserWindow, ipcMain } from 'electron';
-import { platform } from 'os';
-import {
-  mediaLoaded,
-  mediaPlaying,
-  mediaPaused,
-  mediaPlaybackError,
-  mediaLoadError,
-} from '../features/embed/embedSlice';
-import { AppStore, Track, TrackPreviewEmbedSize } from '../common/types';
-import { log } from './helpers/console';
-import { selectTrackSourceByIndex } from '../features/tracks/tracksSlice';
-import { createBrowser } from '../features/browsers/browsersSlice';
-import { sanitiseUrl } from './browser';
+import { platform } from 'node:os';
+
+import { BrowserView, BrowserWindow } from 'electron';
+
+import { EmbedStatus, Track, TrackPreviewEmbedSize } from '../common/types.js';
+import { log } from './helpers/console.js';
+import { selectTrackSourceByIndex } from '../features/tracks/index.js';
+import { loadAndPlayNextTrack } from '../features/embed/index.js';
+import { sanitiseUrl } from './browser.js';
+import { getStore, getDispatch } from './store/index.js';
 // function delay(ms: number) {
 //   return new Promise((resolve) => setTimeout(resolve, ms));
 // }
@@ -21,8 +17,10 @@ let lastClickPlayTime: number;
 let playTimeout: NodeJS.Timeout;
 let loadTimeout: NodeJS.Timeout;
 
-export function initEmbed(mainWindow: BrowserWindow, reduxStore: AppStore): void {
-  const { dispatch, getState } = reduxStore;
+export function initEmbed(mainWindow: BrowserWindow): void {
+  const store = getStore();
+  const dispatch = getDispatch();
+  const { getState } = store;
 
   const setBounds = (embedSize = getState().ui.trackPreviewEmbedSize) => {
     const { height } = mainWindow.getBounds();
@@ -45,7 +43,7 @@ export function initEmbed(mainWindow: BrowserWindow, reduxStore: AppStore): void
   });
 
   embed.webContents.setWindowOpenHandler(({ url }) => {
-    dispatch(createBrowser({ url: sanitiseUrl(url) }));
+    dispatch('BROWSER:CREATE', { url: sanitiseUrl(url) });
     return { action: 'deny' };
   });
 
@@ -76,7 +74,7 @@ export function initEmbed(mainWindow: BrowserWindow, reduxStore: AppStore): void
           void (async () => {
             isPlaying = !!(await embed.webContents.executeJavaScript('$("#player").hasClass("playing");', true));
             if (!isPlaying) {
-              dispatch(mediaPlaybackError());
+              dispatch('EMBED:MEDIA_PLAYBACK_ERROR');
               // TODO: update status panel
             }
           })();
@@ -88,7 +86,7 @@ export function initEmbed(mainWindow: BrowserWindow, reduxStore: AppStore): void
   const mediaStartedPlayingHandler = () => {
     clearTimeout(playTimeout);
     log('playing from embed', Date.now());
-    dispatch(mediaPlaying());
+    dispatch('EMBED:MEDIA_PLAYING');
   };
 
   const mediaPausedHandler = () => {
@@ -108,11 +106,7 @@ export function initEmbed(mainWindow: BrowserWindow, reduxStore: AppStore): void
       }
 
       log('dispatching setPaused, track ended = ', pausedByTrackEnding);
-      dispatch(mediaPaused());
-      if (pausedByTrackEnding) {
-        log('sending autoplay message');
-        mainWindow.webContents.send('handle-autoplay');
-      }
+      dispatch(pausedByTrackEnding ? 'EMBED:MEDIA_PLAYBACK_COMPLETE' : 'EMBED:MEDIA_PAUSED');
 
       log(Date.now() - lastClickPlayTime);
     })();
@@ -121,40 +115,60 @@ export function initEmbed(mainWindow: BrowserWindow, reduxStore: AppStore): void
   embed.webContents.on('media-started-playing', mediaStartedPlayingHandler);
   embed.webContents.on('media-paused', mediaPausedHandler);
 
-  ipcMain.handle('play-track', () => trigger('play'));
-  ipcMain.handle('pause-track', () => trigger('pause'));
-  ipcMain.handle(
-    'load-track',
-    (_event, [trackId]) =>
-      new Promise((resolve) => {
-        if (!trackId) {
-          resolve(false);
-        }
-        const loadFinishedHandler = () => {
-          void (async () => {
-            clearTimeout(loadTimeout);
-            log('loaded', Date.now());
-            dispatch(mediaLoaded());
-            await embed.webContents.executeJavaScript(
-              '$("audio").on("ended", () => { window.playbackComplete = true }).length;',
-              true,
-            );
-            resolve(true);
-          })();
-        };
-        const appState = getState();
-        const { trackPreviewEmbedSize } = appState.ui;
-        embed.webContents.once('did-finish-load', loadFinishedHandler);
+  const loadFinishedHandler = (lastEmbedStatus: EmbedStatus) => {
+    void (async () => {
+      clearTimeout(loadTimeout);
+      log('loaded', Date.now());
+      dispatch('EMBED:MEDIA_LOADED');
+      await embed.webContents.executeJavaScript(
+        '$("audio").on("ended", () => { window.playbackComplete = true }).length;',
+        true,
+      );
+      // ensure previous play status is persisted after the load
+      if (lastEmbedStatus === EmbedStatus.Playing) {
+        dispatch('EMBED:REQUEST_PLAY');
+      }
+    })();
+  };
+
+  let lastEmbedStatus: EmbedStatus;
+  const unsubscribeEmbedListener = store.subscribe((state) => {
+    if (state.embed.status !== lastEmbedStatus) {
+      // invoke actions based on new embed status
+      if (state.embed.status === EmbedStatus.LoadRequested) {
+        const { trackPreviewEmbedSize } = state.ui;
+        const { trackId } = state.embed;
+        embed.webContents.once('did-finish-load', () => loadFinishedHandler(lastEmbedStatus));
         const trackSourceSelector = selectTrackSourceByIndex(trackId as Track['id'], 0);
-        const trackSource = trackSourceSelector(appState);
-        const embedSize = trackPreviewEmbedSize.toLowerCase();
-        const trackUrl = `https://bandcamp.com/EmbeddedPlayer/size=${embedSize}/bgcol=ffffff/linkcol=0687f5/track=${trackSource.sourceId}/transparent=true/`;
-        void embed.webContents.loadURL(trackUrl);
-        loadTimeout = setTimeout(() => {
-          dispatch(mediaLoadError());
-          // TODO: update status panel
-        }, 30000);
+        const trackSource = trackSourceSelector(state);
+        if (trackSource) {
+          const embedSize = trackPreviewEmbedSize.toLowerCase();
+          const trackUrl = `https://bandcamp.com/EmbeddedPlayer/size=${embedSize}/bgcol=ffffff/linkcol=0687f5/track=${trackSource.sourceId}/transparent=true/`;
+          void embed.webContents.loadURL(trackUrl);
+          loadTimeout = setTimeout(() => {
+            dispatch('EMBED:MEDIA_LOAD_ERROR');
+            // TODO: update status panel
+          }, 30000);
+        }
         setBounds(trackPreviewEmbedSize);
-      }),
-  );
+        // await window.api.invoke('resize-browsers');
+
+        // embed status has changed
+        lastEmbedStatus = state.embed.status;
+      } else if (state.embed.status === EmbedStatus.PlayRequested) {
+        trigger('play');
+      } else if (state.embed.status === EmbedStatus.PauseRequested) {
+        trigger('pause');
+      } else if (state.embed.status === EmbedStatus.PlaybackComplete) {
+        const { trackId, autoplayEnabled } = state.embed;
+        log('handleAutoplay', autoplayEnabled, trackId);
+        if (autoplayEnabled && trackId) {
+          // get next track
+          dispatch(loadAndPlayNextTrack());
+        }
+      }
+    }
+  });
+
+  mainWindow.on('close', unsubscribeEmbedListener);
 }
