@@ -1,17 +1,18 @@
-import { Store } from '@reduxjs/toolkit';
 import { IpcRenderer } from 'electron';
-import React from 'react';
-import { Provider } from 'react-redux';
-import * as ReactDOM from 'react-dom';
+import React, { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
-import { syncRenderer } from '@goosewobbler/electron-redux/renderer';
-import { App } from './App';
-import { createReduxStore } from '../common/reduxStore';
-import '../css/tailwind.src.pcss';
+
+import { App } from './App.js';
+import type { PreloadZustandBridgeReturn } from 'zutron';
+import type { AppState } from '../common/types.js';
+
+import '../index.css';
 
 declare global {
   interface Window {
+    zutron: PreloadZustandBridgeReturn<AppState>['handlers'];
     api: {
       isDev: boolean;
       invoke(channel: string, ...data: unknown[]): Promise<unknown>;
@@ -21,24 +22,38 @@ declare global {
   }
 }
 
-function render(reduxStore: Store): void {
-  ReactDOM.render(
-    <Provider store={reduxStore}>
+// function render(reduxStore: Store): void {
+//   ReactDOM.render(
+//     <Provider store={reduxStore}>
+//       <DndProvider backend={HTML5Backend}>
+//         <App />
+//       </DndProvider>
+//     </Provider>,
+//     document.getElementById('app'),
+//   );
+// }
+
+function render(): void {
+  const container = document.getElementById('app')!;
+  const root = createRoot(container);
+  root.render(
+    <StrictMode>
       <DndProvider backend={HTML5Backend}>
         <App />
       </DndProvider>
-    </Provider>,
-    document.getElementById('app'),
+    </StrictMode>,
   );
 }
 
-void (async () => {
-  const reduxStore = await createReduxStore({
-    context: 'renderer',
-    syncFn: syncRenderer,
-  });
-  render(reduxStore);
-})();
+render();
+
+// void (async () => {
+//   const reduxStore = await createReduxStore({
+//     context: 'renderer',
+//     // syncFn: syncRenderer,
+//   });
+//   render(reduxStore);
+// })();
 
 // if (isDev) {
 //   // eslint-disable-next-line
